Split WAL frame encoding into a module
Move frame serialization, deserialization, scanning, and checksum helpers out of the log manager.

The log manager keeps ownership and durability state while the frame module owns the binary WAL format details.
